Retrying mechanism added.
