+++ src/main/java/com/example/projectfirst/connector/ConnectorService.java
@@ -1,9 +1,10 @@
 package com.example.projectfirst.connector;
 
+import com.example.projectfirst.connector.exception.ConnectorNotFoundException;
+import com.example.projectfirst.connector.model.Connector;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
-import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import java.io.IOException;
@@ -12,7 +13,7 @@ import java.util.List;
 import java.util.Map;
 
 @Service
-public class ConnectorService {
+public class ConnectorService implements ConnectorInterface{
 
     @Autowired
     private ConnectorRepository connectorRepository;
