+++ src/main/java/com/example/projectfirst/connector/model/Spec.java
@@ -1,4 +1,4 @@
-package com.example.projectfirst.connector;
+package com.example.projectfirst.connector.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonTypeName;
