+++ src/main/java/com/example/projectfirst/pipeline/apiRequestHandler/StepHandler.java
@@ -1,8 +1,8 @@
-package com.example.projectfirst.pipeline;
+package com.example.projectfirst.pipeline.apiRequestHandler;
 
 import com.example.projectfirst.connector.ConnectorService;
+import com.example.projectfirst.pipeline.model.StepParameters;
 import com.squareup.okhttp.Response;
-import org.springframework.stereotype.Service;
 
 import java.io.IOException;
