+++ src/main/java/com/example/projectfirst/pipeline/model/SpecPost.java
@@ -1,5 +1,6 @@
-package com.example.projectfirst.pipeline;
+package com.example.projectfirst.pipeline.model;
 
+import com.example.projectfirst.pipeline.model.Spec;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import lombok.*;
 import lombok.extern.jackson.Jacksonized;
@@ -7,9 +8,9 @@ import lombok.extern.jackson.Jacksonized;
 @Value
 @Jacksonized
 @Builder
-@JsonTypeName("API_GET")
-public class SpecGet implements Spec{
-    String output;
+@JsonTypeName("API_POST")
+public class SpecPost implements Spec {
+    String body;
     String url;
     String connectorID;
 }
