+++ src/main/java/com/example/projectfirst/pipelineExecution/PipelineExecutionInterface.java
@@ -1,0 +1,13 @@
+package com.example.projectfirst.pipelineExecution;
+
+import java.util.List;
+
+public interface PipelineExecutionInterface {
+    List<PipelineExecutionCollection> fetchAllExecutions();
+    PipelineExecutionCollection fetchExecution(String id);
+    List<PipelineExecutionCollection> fetchPausedExecutions();
+    String executePipeline(String id);
+    String resumeExecution(String id);
+    String deleteExecution(String id);
+    void deleteExecutions();
+}
