+++ src/main/java/com/example/projectfirst/pipelineExecution/exception/PipelineExecutionNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.projectfirst.pipelineExecution;
+package com.example.projectfirst.pipelineExecution.exception;
 
 public class PipelineExecutionNotFoundException extends RuntimeException{
     public PipelineExecutionNotFoundException(String id){
