+++ src/main/java/com/example/projectfirst/pipelineExecution/services/StateService.java
@@ -1,5 +1,8 @@
-package com.example.projectfirst.pipelineExecution;
+package com.example.projectfirst.pipelineExecution.services;
 
+import com.example.projectfirst.pipelineExecution.PipelineExecutionCollection;
+import com.example.projectfirst.pipelineExecution.PipelineExecutionRepository;
+import com.example.projectfirst.pipelineExecution.exception.PipelineExecutionNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,12 +17,18 @@ public class StateService {
         Optional<PipelineExecutionCollection> pipelineExecution =
                 pipelineExecutionRepository.findById(pipelineExeId);
 
+        if(pipelineExecution.isEmpty())
+            throw new PipelineExecutionNotFoundException(pipelineExeId);
+
         return pipelineExecution.get().getState();
     }
 
     public void setState(String pipelineExeId, String state) {
         Optional<PipelineExecutionCollection> pipelineExecution =
                 pipelineExecutionRepository.findById(pipelineExeId);
+
+        if(pipelineExecution.isEmpty())
+            throw new PipelineExecutionNotFoundException(pipelineExeId);
 
         pipelineExecution.get().setState(state);
         pipelineExecutionRepository.save(pipelineExecution.get());
